adding wait time algorithm

+++ Models/WaitTime.cs
@@ -3,6 +3,6 @@ namespace API.Models
     public class WaitTime
     {
         public string status{get;set;}
-        public int waitTime{get;set;}
+        public float waitTime{get;set;}
     }
 }
